requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.6
h11==0.14.0
idna==3.10
iniconfig==2.0.0
-e git+https://github.com/chetbae/kilterboard-heatmap-backend.git@e6114e2b1653b942c02130378074b1da8871e6fc#egg=kilterboard_heatmap
lz4==4.3.3
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
pydantic==2.10.5
pydantic_core==2.27.2
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
sqlitecloud==0.0.83
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
